+++ weblate/trans/tests/test_search.py
@@ -301,6 +301,14 @@ class BulkStateTest(ViewTestCase):
         self.assertContains(response, "Bulk edit completed, 1 string was updated.")
         unit = self.get_unit()
         self.assertTrue(label in unit.labels.all())
+        self.assertEqual(
+            getattr(unit.translation.stats, "label:{}".format(label.name)), 1
+        )
+        unit.source_info.translation.stats.invalidate()
+        self.assertEqual(
+            getattr(unit.source_info.translation.stats, "label:{}".format(label.name)),
+            1,
+        )
         response = self.client.post(
             reverse("bulk-edit", kwargs=self.kw_project),
             {"q": "state:needs-editing", "state": -1, "remove_labels": label.pk},
@@ -309,3 +317,11 @@ class BulkStateTest(ViewTestCase):
         self.assertContains(response, "Bulk edit completed, 1 string was updated.")
         unit = self.get_unit()
         self.assertFalse(label in unit.labels.all())
+        self.assertEqual(
+            getattr(unit.translation.stats, "label:{}".format(label.name)), 0
+        )
+        unit.source_info.translation.stats.invalidate()
+        self.assertEqual(
+            getattr(unit.source_info.translation.stats, "label:{}".format(label.name)),
+            0,
+        )